requirements.txt
Flask==3.0.3
reportlab==4.2.5
Pillow==10.4.0
pywebview>=5.0,<6
